moving and renaming files to follow the LLVM conventions
