+++ lib/fir/Dialect/CMakeLists.txt
@@ -1,0 +1,29 @@
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-error")
+
+add_llvm_library(FIRDialect
+  FIRAttr.cpp
+  FIRDialect.cpp
+  FIROps.cpp
+  FIRType.cpp
+)
+
+add_dependencies(FIRDialect FIROpsIncGen)
+
+target_link_libraries(FIRDialect
+  MLIRTargetLLVMIR
+  MLIRTargetLLVMIRModuleTranslation
+  MLIREDSC
+  MLIRExecutionEngine
+  MLIRParser
+  MLIRSupport
+  MLIRStandardToLLVM
+  MLIRTransforms
+  LLVMAsmParser
+  LLVMAsmPrinter
+  LLVMRemarks
+)
+
+install (TARGETS FIRDialect
+  ARCHIVE DESTINATION lib
+  LIBRARY DESTINATION lib
+)
