+++ lib/fir/Dialect/FIRDialect.cpp
@@ -1,4 +1,4 @@
-//===-- lib/optimizer/FIRDialect.cpp ----------------------------*- C++ -*-===//
+//===-- fir/FIRDialect.cpp ------------------------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,13 +6,13 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "optimizer/FIRDialect.h"
+#include "fir/Dialect/FIRDialect.h"
+#include "fir/Dialect/FIRAttr.h"
+#include "fir/Dialect/FIROps.h"
+#include "fir/Dialect/FIRType.h"
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
 #include "mlir/IR/StandardTypes.h"
 #include "mlir/Transforms/SideEffectsInterface.h"
-#include "optimizer/Attribute.h"
-#include "optimizer/FIROps.h"
-#include "optimizer/FIRType.h"
 
 namespace M = mlir;
