+++ lib/lower/CMakeLists.txt
@@ -1,23 +1,15 @@
-#===-- lib/lower/CMakeLists.txt --------------------------------------------===#
-#
-# Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-# See https://llvm.org/LICENSE.txt for license information.
-# SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-#
-#===----------------------------------------------------------------------===--#
-
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wno-error")
 
 add_library(FortranLower
-  ast-builder.cc
-  bridge.cc
-  builder.cc
-  convert-expr.cc
-  convert-type.cc
-  intrinsics.cc
-  mangler.cc
-  runtime.cc
-  io.cc
+  ASTBuilder.cpp
+  Bridge.cpp
+  OpBuilder.cpp
+  ConvertExpr.cpp
+  ConvertType.cpp
+  Intrinsics.cpp
+  Mangler.cpp
+  Runtime.cpp
+  IO.cpp
 )
 
 target_link_libraries(FortranLower
