+++ lib/lower/ConvertExpr.cpp
@@ -6,7 +6,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "convert-expr.h"
+#include "flang/lower/ConvertExpr.h"
 #include "../common/default-kinds.h"
 #include "../common/unwrap.h"
 #include "../evaluate/fold.h"
@@ -14,10 +14,14 @@
 #include "../semantics/expression.h"
 #include "../semantics/symbol.h"
 #include "../semantics/type.h"
-#include "bridge.h"
-#include "builder.h"
-#include "complex-handler.h"
-#include "convert-type.h"
+#include "fir/Dialect/FIRDialect.h"
+#include "fir/Dialect/FIROps.h"
+#include "fir/Dialect/FIRType.h"
+#include "flang/lower/Bridge.h"
+#include "flang/lower/ComplexHandler.h"
+#include "flang/lower/ConvertType.h"
+#include "flang/lower/OpBuilder.h"
+#include "flang/lower/Runtime.h"
 #include "mlir/Dialect/AffineOps/AffineOps.h"
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
 #include "mlir/Dialect/StandardOps/Ops.h"
@@ -28,10 +32,6 @@
 #include "mlir/Pass/PassManager.h"
 #include "mlir/Transforms/DialectConversion.h"
 #include "mlir/Transforms/Passes.h"
-#include "optimizer/FIRDialect.h"
-#include "optimizer/FIROps.h"
-#include "optimizer/FIRType.h"
-#include "runtime.h"
 #include "llvm/ADT/APFloat.h"
 #include "llvm/ADT/ArrayRef.h"
 #include "llvm/IR/Type.h"
