+++ lib/lower/Intrinsics.cpp
@@ -6,13 +6,13 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "intrinsics.h"
-#include "builder.h"
-#include "complex-handler.h"
-#include "convert-type.h"
-#include "optimizer/FIROps.h"
-#include "optimizer/FIRType.h"
-#include "runtime.h"
+#include "flang/lower/Intrinsics.h"
+#include "fir/Dialect/FIROps.h"
+#include "fir/Dialect/FIRType.h"
+#include "flang/lower/ComplexHandler.h"
+#include "flang/lower/ConvertType.h"
+#include "flang/lower/OpBuilder.h"
+#include "flang/lower/Runtime.h"
 #include "llvm/ADT/Optional.h"
 #include "llvm/ADT/Twine.h"
 #include <algorithm>
