+++ lib/lower/Runtime.cpp
@@ -1,4 +1,4 @@
-//===-- lib/lower/runtime.cc ----------------------------------------------===//
+//===-- lower/Runtime.cpp -------------------------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,11 +6,11 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "runtime.h"
-#include "builder.h"
+#include "flang/lower/Runtime.h"
+#include "fir/Dialect/FIRType.h"
+#include "flang/lower/OpBuilder.h"
 #include "mlir/IR/StandardTypes.h"
 #include "mlir/IR/Types.h"
-#include "optimizer/FIRType.h"
 #include "llvm/ADT/ArrayRef.h"
 #include "llvm/ADT/SmallVector.h"
 #include "llvm/ADT/StringRef.h"
