+++ lib/optimizer/IteratedDominanceFrontier.cpp
@@ -10,7 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "optimizer/Analysis/IteratedDominanceFrontier.h"
+#include "flang/optimizer/Analysis/IteratedDominanceFrontier.h"
 #include "mlir/Analysis/Dominance.h"
 #include "llvm/ADT/DenseMap.h"
 #include "llvm/ADT/SmallPtrSet.h"
