+++ lib/optimizer/KindMapping.cpp
@@ -6,7 +6,7 @@
 //
 //===----------------------------------------------------------------------===//
 
-#include "optimizer/KindMapping.h"
+#include "flang/optimizer/KindMapping.h"
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
 #include "llvm/ADT/Optional.h"
 #include "llvm/Support/CommandLine.h"
