+++ lib/optimizer/Transforms/CSE.cpp
@@ -12,6 +12,8 @@
 ///
 //===----------------------------------------------------------------------===//
 
+#include "fir/Dialect/FIROpsSupport.h"
+#include "flang/optimizer/Transforms/Passes.h"
 #include "mlir/Analysis/Dominance.h"
 #include "mlir/IR/Attributes.h"
 #include "mlir/IR/Builders.h"
@@ -20,8 +22,6 @@
 #include "mlir/Support/Functional.h"
 #include "mlir/Transforms/Passes.h"
 #include "mlir/Transforms/Utils.h"
-#include "optimizer/FIROpsSupport.h"
-#include "optimizer/Transforms/Passes.h"
 #include "llvm/ADT/DenseMapInfo.h"
 #include "llvm/ADT/Hashing.h"
 #include "llvm/ADT/ScopedHashTable.h"
