+++ tools/bbc/bbc.cpp
@@ -13,8 +13,6 @@
 
 #include "../../lib/common/Fortran-features.h"
 #include "../../lib/common/default-kinds.h"
-#include "../../lib/lower/bridge.h"
-#include "../../lib/lower/convert-expr.h"
 #include "../../lib/parser/characters.h"
 #include "../../lib/parser/dump-parse-tree.h"
 #include "../../lib/parser/message.h"
@@ -26,6 +24,14 @@
 #include "../../lib/semantics/expression.h"
 #include "../../lib/semantics/semantics.h"
 #include "../../lib/semantics/unparse-with-symbols.h"
+#include "fir/Dialect/FIRDialect.h"
+#include "flang/lower/Bridge.h"
+#include "flang/lower/ConvertExpr.h"
+#include "flang/optimizer/CodeGen/CodeGen.h"
+#include "flang/optimizer/InternalNames.h"
+#include "flang/optimizer/KindMapping.h"
+#include "flang/optimizer/Transforms/Passes.h"
+#include "flang/optimizer/Transforms/StdConverter.h"
 #include "mlir/Conversion/LoopToStandard/ConvertLoopToStandard.h"
 #include "mlir/IR/MLIRContext.h"
 #include "mlir/IR/Module.h"
@@ -33,12 +39,6 @@
 #include "mlir/Pass/Pass.h"
 #include "mlir/Pass/PassManager.h"
 #include "mlir/Transforms/Passes.h"
-#include "optimizer/CodeGen/CodeGen.h"
-#include "optimizer/FIRDialect.h"
-#include "optimizer/InternalNames.h"
-#include "optimizer/KindMapping.h"
-#include "optimizer/Transforms/Passes.h"
-#include "optimizer/Transforms/StdConverter.h"
 #include "llvm/Support/CommandLine.h"
 #include "llvm/Support/ErrorOr.h"
 #include "llvm/Support/InitLLVM.h"
